Allow user to change hoonymode winning score through fraglimit, & round length through timelimit.
If timelimit is hit and fraglimit hasn't been reached, both players get
a frag, this keeps game advancing towards an end.

Being ahead by 1 frag is enough after the end of each round of spawns.
This should have no effect when no timelimit is in place, but should lead
to shorter games by essentially having sudden death overtime, but on a
round-of-spawns basis.